Split tests into accuracy and performance. And started performance tests
