+++ README.md
@@ -152,3 +152,5 @@ Using the [gh-pages](https://www.npmjs.com/package/gh-pages) for debugger/demo d
 * [Google Material Icons](https://material.io/icons/)
 
 * [The Cycle Accurate Game Boy Docs](https://github.com/AntonioND/giibiiadvance/blob/master/docs/TCAGBD.pdf)
+
+* [Demos for perf testing GB/GBC](http://privat.bahnhof.se/wb800787/gb/demos/)
