+++ test/accuracy/accuracy-test.js
@@ -1,20 +1,14 @@
-// Assertion
-const assert = require('assert');
+// Common test functions
+const commonTest = require('../common-test');
 
 // Wasm Boy library
-const WasmBoy = require('../dist/wasmboy.cjs.js').WasmBoy;
+const WasmBoy = require('../../dist/wasmboy.cjs.js').WasmBoy;
 
 // File management
 const fs = require('fs');
-const path = require('path');
 
-// Image Creation
-const PNGImage = require('pngjs-image');
-
-
-// Define some constants
-const GAMEBOY_CAMERA_WIDTH = 160;
-const GAMEBOY_CAMERA_HEIGHT = 144;
+// Assertion
+const assert = require('assert');
 
 // Some Timeouts for specified test roms
 // Default is 20 seconds, as it runs cpu_instrs in that time
@@ -31,54 +25,11 @@ WasmBoy.initialize(false, {
     isGbcEnabled: true
 });
 
-// Function to create an image from output
-const createImageFromFrame = (imageDataArray, outputPath) => {
-  return new Promise((resolve, reject) => {
-    // https://www.npmjs.com/package/pngjs-image
-    const image = PNGImage.createImage(GAMEBOY_CAMERA_WIDTH, GAMEBOY_CAMERA_HEIGHT);
+const testRomsPath = './test/accuracy/testroms';
 
-    // Write our pixel values
-    for (let i = 0; i < imageDataArray.length - 4; i = i + 4) {
-
-      // Since 4 indexes represent 1 pixels. divide i by 4
-      const pixelIndex = i / 4;
-
-      // Get our y value from i
-      const y = Math.floor(pixelIndex / GAMEBOY_CAMERA_WIDTH);
-
-      // Get our x value from i
-      const x = pixelIndex % GAMEBOY_CAMERA_WIDTH;
-
-      image.setAt(x, y, {
-        red: imageDataArray[i],
-        green: imageDataArray[i + 1],
-        blue: imageDataArray[i + 2],
-        alpha: imageDataArray[i + 3],
-      });
-    }
-
-    image.writeImage(outputPath, function (err) {
-        if (err) {
-          reject(err);
-        }
-        resolve();
-    });
-  });
-}
-
-// Get our folders under testroms
-const isDirectory = source => fs.lstatSync(source).isDirectory()
-const getDirectories = source =>
-  fs.readdirSync(source).map(name => path.join(source, name)).filter(isDirectory);
-
-const testRomsPath = './test/testroms';
-
-getDirectories(testRomsPath).forEach((directory) => {
+commonTest.getDirectories(testRomsPath).forEach((directory) => {
   // Get all test roms for the directory
-  const files = fs.readdirSync(directory);
-  const testRoms = files.filter(function(file) {
-      return path.extname(file).toLowerCase() === '.gb';
-  });
+  const testRoms = commonTest.getAllRomsInDirectory(directory);
 
   // Create a describe for the directory
   describe(directory, () => {
@@ -126,38 +77,11 @@ getDirectories(testRomsPath).forEach((directory) => {
             WasmBoy.pauseGame().then(() => {
               console.log(`Checking results for the following test rom: ${directory}/${testRom}`);
 
-              // Going to compare pixel values from the VRAM to confirm tests
-              const imageDataArray = [];
-              const rgbColor = [];
-
-              for(let y = 0; y < GAMEBOY_CAMERA_HEIGHT; y++) {
-                for (let x = 0; x < GAMEBOY_CAMERA_WIDTH; x++) {
-
-                  // Each color has an R G B component
-                  let pixelStart = ((y * 160) + x) * 3;
-
-                  for (let color = 0; color < 3; color++) {
-                    rgbColor[color] = WasmBoy.wasmByteMemory[
-                      WasmBoy.wasmInstance.exports.frameInProgressVideoOutputLocation + pixelStart + color
-                    ];
-                  }
-
-                  // Doing graphics using second answer on:
-                  // https://stackoverflow.com/questions/4899799/whats-the-best-way-to-set-a-single-pixel-in-an-html5-canvas
-                  // Image Data mapping
-                  const imageDataIndex = (x + (y * GAMEBOY_CAMERA_WIDTH)) * 4;
-
-                  imageDataArray[imageDataIndex] = rgbColor[0];
-                  imageDataArray[imageDataIndex + 1] = rgbColor[1];
-                  imageDataArray[imageDataIndex + 2] = rgbColor[2];
-                  // Alpha, no transparency
-                  imageDataArray[imageDataIndex + 3] = 255;
-                }
-              }
+              const imageDataArray = commonTest.getImageDataFromFrame();
 
               // Output a gitignored image of the current tests
               const testImagePath = testRom.replace('.gb', '.current.png');
-              createImageFromFrame(imageDataArray, `${directory}/${testImagePath}`).then(() => {
+              commonTest.createImageFromFrame(imageDataArray, `${directory}/${testImagePath}`).then(() => {
                 // Now compare with the current array if we have it
                 const testDataPath = testRom.replace('.gb', '.golden.output');
                 if (fs.existsSync(`${directory}/${testDataPath}`)) {
@@ -203,7 +127,7 @@ getDirectories(testRomsPath).forEach((directory) => {
                   fs.writeFileSync(`${directory}/${testDataPath}`, imageDataStringified);
 
                   const testImagePath = testRom.replace('.gb', '.golden.png');
-                  createImageFromFrame(imageDataArray, `${directory}/${testImagePath}`).then(() => {
+                  commonTest.createImageFromFrame(imageDataArray, `${directory}/${testImagePath}`).then(() => {
                     done();
                   });
                 }
