+++ test/accuracy/testroms/blargg/cpu_instrs.current.png
