+++ test/performance/performance-test.js
@@ -1,0 +1,2 @@
+// Wasm Boy library
+const WasmBoy = require('../../dist/wasmboy.cjs.js').WasmBoy;
