+++ test/performance/testroms/README.md
@@ -1,0 +1,87 @@
+Back to Color
+=============
+
+Introduction
+------------
+
+Hello!
+
+This time I'm here to show you my first GBC demo! Specially made for the
+[2014 GBDev Game Boy Coding Compo](http://gbdev.gg8.se/forums/viewtopic.php?id=173).
+
+I started it like four years ago, but stopped after doing 3 scenes (the first
+effect, the text scroller at the end of the first song and the colorful horizontal
+bars before the matrix effect). I've coded the rest in just a few weeks, in fact
+I started doing new effects a few days before the GBDev compo announcement. I've
+coded it and made the music, the only thing I haven't done is the train screen
+graphics and the credits graphics. I'm not a very good musician, so... Oh, and it
+shows a little credits screen in non-color GBs!
+
+If you want to test it in an emulator use one of:
+- [BGB](http://bgb.bircd.org/)
+- [Gambatte](http://sourceforge.net/projects/gambatte/)
+- [VBA-M](http://vba-m.com/)
+- My [GiiBii Advance](https://github.com/AntonioND/giibiiadvance) :P
+
+Most emulators have a "Frame Blending" option, enable it to simulate the real
+Game Boy LCD. It works in real hardware, of course! I've tested it in DMG, MGB,
+GBC, GBA and GBA SP. Non-color GB only show the reduced credits screen!
+
+If you want to skip some parts, hold up/right/down/left when the demo is loading.
+- Up: Go to the effect before the blobs.
+- Right: Go to the pentagon effect.
+- Down: Go to the fire effect.
+- Left: Credits.
+
+The demo has been made using:
+- [My fork of RGBDS](https://github.com/AntonioND/rgbds) (with a few bugfixes)
+- A [GBT Player](https://github.com/AntonioND/gbt-player) version modified to use
+  only the first ROM bank.
+- OpenMPT to make the music.
+- GBTK and GBMB.
+- MSPaint :P
+- A bit of Photoshop
+
+I hope you like it! :)
+
+[GitHub repository](https://github.com/AntonioND/back-to-color)
+
+[YouTube video](https://www.youtube.com/watch?v=eSEsd1OGA74)
+
+Credits
+-------
+
+* Code, music, most graphics: Antonio Niño Díaz (AntonioND)
+  * Emails: antonio_nd@outlook.com / antonionidi@gmail.com
+  * Webpage: [http://antoniond_blog.drunkencoders.com/](http://antoniond_blog.drunkencoders.com/)
+  * Old webpage: [http://antoniond.drunkencoders.com/](http://antoniond.drunkencoders.com/)
+* Train and credits art: jordirafel
+  * Email: jordirafel@gmail.com
+
+
+License (BSD)
+-------------
+
+Copyright (c) 2014, Antonio Niño Díaz (AntonioND)
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+* Redistributions of source code must retain the above copyright notice, this
+  list of conditions and the following disclaimer.
+
+* Redistributions in binary form must reproduce the above copyright notice,
+  this list of conditions and the following disclaimer in the documentation
+  and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
